+++ lib/php/test/EmojiTest.php
@@ -9,7 +9,7 @@ use JoyPixels\JoyPixels;
  */
 class EmojiTest extends \PHPUnit_Framework_TestCase
 {
-    private $emojiVersion = '4.5';
+    private $emojiVersion = '5.0';
 
     public function emojiProvider()
     {