+++ lib/php/test/JoyPixelsTest.php
@@ -7,7 +7,7 @@ use JoyPixels\JoyPixels;
 class JoyPixelsTest extends \PHPUnit_Framework_TestCase
 {
 
-    private $emojiVersion = '4.5';
+    private $emojiVersion = '5.0';
 
     public function emojiProvider()
     {